requirements.txt
reportlab
markdown
prompt_toolkit
pypdf
pillow
requests
weasyprint
rich
sympy
